Solved bug causing the app to crush after location permission granted through settings window
The app was crushing when the user was denying initially location
permission during first intallation and was granting it implicitly via
the Observe or Prediction panel when trying to submit an observation or
prediction. The app on resume was  fast requesting location updates
without the location manager being ready. I extended the condition in
locationAvailable() to include the criterion of location manager being
initialised before any further action